Fix vulnerabilities W1-W7 and X1-X8, add doc/FIXES_CLAUDE_20260610.md
Remediation of the 2026-05-29 (W1-W7) and 2026-06-10 (X1-X8) audit
findings:
- W1: guard m_dev/m_thread in AirspyHFSource::stop()
- W2: validate hf_att range before narrowing to uint8_t
- W3: enforce Airspy HF frequency lower bound of 1 kHz
- W4/X3: bounds-check sampleRateIndex in private configure()
- W5: null-check m_buf in Airspy/AirspyHF rx callbacks
- W6: size_t arithmetic in channel-expansion helpers
- W7: fclose PPS output file at exit
- X1: cap blklen at IfResampler max input length (65536)
- X2: reject |ppm| >= 1000000 and non-positive effective ifrate;
  add 64x max IF-upsampling-ratio guard
- X4: AirspySource::start() returns true, not *this
- X5: make PortAudioOutput::output_close() idempotent and guarded
- X6: report and exit on audio_output->write() failure
- X7: bound DataBuffer queue at 1024 blocks, drop-oldest with warning
- X8: guard ppm display against zero frequency

Details in doc/FIXES_CLAUDE_20260610.md.

+++ sfmbase/FmDecode.cpp
@@ -241,10 +241,10 @@ inline void FmDecoder::demod_stereo(const SampleVector &samples_baseband,
 // Duplicate mono signal in left/right channels.
 inline void FmDecoder::mono_to_left_right(const SampleVector &samples_mono,
                                           SampleVector &audio) {
-  unsigned int n = samples_mono.size();
+  size_t n = samples_mono.size();
 
   audio.resize(2 * n);
-  for (unsigned int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     Sample m = samples_mono[i];
     audio[2 * i] = m;
     audio[2 * i + 1] = m;
@@ -255,11 +255,11 @@ inline void FmDecoder::mono_to_left_right(const SampleVector &samples_mono,
 inline void FmDecoder::stereo_to_left_right(const SampleVector &samples_mono,
                                             const SampleVector &samples_stereo,
                                             SampleVector &audio) {
-  unsigned int n = samples_mono.size();
+  size_t n = samples_mono.size();
   assert(n == samples_stereo.size());
 
   audio.resize(2 * n);
-  for (unsigned int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     Sample m = samples_mono[i];
     // L-R singal is boosted by 1.017
     // for better separation (suggested by Teruhiko Hayashi)
@@ -273,10 +273,10 @@ inline void FmDecoder::stereo_to_left_right(const SampleVector &samples_mono,
 // (samples_mono used for the size determination only)
 inline void FmDecoder::zero_to_left_right(const SampleVector &samples_mono,
                                           SampleVector &audio) {
-  unsigned int n = samples_mono.size();
+  size_t n = samples_mono.size();
 
   audio.resize(2 * n);
-  for (unsigned int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     audio[2 * i] = 0.0;
     audio[2 * i + 1] = 0.0;
   }
